Rename challsrv -> challtestsrv
