+++ test/challtestsrv/challenge-servers.go
@@ -1,4 +1,4 @@
-package challsrv
+package challtestsrv
 
 import (
 	"fmt"
