+++ test/chisel2.py
@@ -42,7 +42,7 @@ PORT = os.getenv('PORT', '5002')
 
 os.environ.setdefault('REQUESTS_CA_BUNDLE', 'test/wfe-tls/minica.pem')
 
-# URLs for management interface of challsrv
+# URLs for management interface of challtestsrv
 SET_TXT = "http://localhost:8056/set-txt"
 CLEAR_TXT = "http://localhost:8056/clear-txt"
